Fix hanging workers when an unresponsive worker is being killed at the very end of the crawl

+++ lib/src/worker/pool.dart
@@ -17,30 +17,37 @@ class Pool {
   /// The number of threads.
   final int count;
 
+  bool _isShuttingDown = false;
   List<Worker> _workers;
 
   Timer _healthCheckTimer;
-  Map<Worker, DateTime> _lastJobPosted = new Map<Worker, DateTime>();
 
+  Map<Worker, DateTime> _lastJobPosted = new Map<Worker, DateTime>();
   final Set<String> _hostGlobs;
 
   StreamController<FetchResults> _fetchResultsSink =
       new StreamController<FetchResults>();
-  Stream<FetchResults> fetchResults;
 
+  Stream<FetchResults> fetchResults;
   StreamController<String> _messagesSink = new StreamController<String>();
+
   Stream<String> messages;
+  bool _finished = false;
 
   Pool(this.count, this._hostGlobs) {
     fetchResults = _fetchResultsSink.stream;
     messages = _messagesSink.stream;
   }
 
-  bool get allIdle =>
-      _workers.every((worker) => worker.idle || !worker.spawned);
+  ///
+  bool get allIdle => _workers
+      .every((worker) => worker.idle || !worker.spawned || worker.isKilled);
 
-  bool get allWorking =>
-      _workers.every((worker) => !worker.idle || !worker.spawned);
+  bool get anyIdle => _workers.any((worker) => worker.idle);
+
+  bool get finished => _finished;
+
+  bool get isShuttingDown => _isShuttingDown;
 
   Worker check(Destination destination) {
     var worker = pickWorker();
@@ -50,14 +57,12 @@ class Pool {
     return worker;
   }
 
-  bool _finished = false;
-  bool get finished => _finished;
-
   Future<Null> close() async {
+    _isShuttingDown = true;
     _healthCheckTimer.cancel();
     await Future.wait(_workers.map((worker) async {
-      if (worker.isKilled) return;
-      worker.kill();
+      if (!worker.spawned || worker.isKilled) return;
+      await worker.kill();
     }));
     _finished = true;
   }
@@ -94,30 +99,38 @@ class Pool {
     _addHostGlobs();
 
     _healthCheckTimer = new Timer.periodic(healthCheckFrequency, (_) async {
+      if (_isShuttingDown) return null;
       var now = new DateTime.now();
       for (int i = 0; i < _workers.length; i++) {
         var worker = _workers[i];
         if (!worker.idle &&
+            !worker.isKilled &&
             _lastJobPosted[worker] != null &&
             now.difference(_lastJobPosted[worker]) > workerTimeout) {
           _messagesSink.add("Killing unresponsive $worker");
           var destination = worker.destinationToCheck;
+          _lastJobPosted.remove(worker);
+          var newWorker = new Worker()..name = '$i';
+          _workers[i] = newWorker;
+
+          // Only notify about the failed destination when the old
+          // worker is gone. Otherwise, crawl could fail to wrap up, thinking
+          // that one Worker is still working.
           var checked = new DestinationResult.fromDestination(destination);
           checked.didNotConnect = true;
           var result = new FetchResults(checked, const []);
           _fetchResultsSink.add(result);
-          _killUnresponsive(worker);
-          var newWorker = new Worker()..name = '$i';
-          _workers[i] = newWorker;
+
           await newWorker.spawn();
+          if (_isShuttingDown) {
+            // The pool has been closed in the time it took to spawn this
+            // worker.
+            await newWorker.kill();
+          }
+          await worker.kill();
         }
       }
     });
-  }
-
-  void _killUnresponsive(Worker worker) {
-    worker.kill();
-    _lastJobPosted.remove(worker);
   }
 
   /// Sends host globs (e.g. http://example.com/**) to all the workers.
